requirements.txt
streamlit
google-generativeai
python-dotenv
gtts
SpeechRecognition
pyaudio